pyproject.toml
[tool.poetry]
name = "pogodynka"
version = "0.1.0"
description = "Pogodynka collecting data from temperature and PM sensors"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "3.7.3"
google-cloud-bigquery = "^2.17.0"
pyserial = "^3.5"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
pre-commit = "^2.13.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
